requirements.txt
pycryptodome
Pillow
tkinter
